+++ benchmarks/pvc/pvc_gemm_bf16_bf16_fp32.cpp
@@ -32,8 +32,8 @@
 #pragma once
 
 #include "cutlass/epilogue/collective/default_epilogue.hpp"
-#include "cutlass/epilogue/collective/intel_pvc_epilogue.hpp"
-#include "cutlass/epilogue/fusion/intel_pvc_callbacks.hpp"
+#include "cutlass/epilogue/collective/xe_epilogue.hpp"
+#include "cutlass/epilogue/fusion/xe_callbacks.hpp"
 
 using namespace cute;
