+++ include/cutlass/epilogue/collective/builders/xe_builder.inl
@@ -33,8 +33,8 @@
 
 #include <cutlass/arch/arch.h>
 #include "cutlass/epilogue/collective/default_epilogue.hpp"
-#include "cutlass/epilogue/collective/intel_pvc_epilogue.hpp"
-#include "cutlass/epilogue/fusion/intel_pvc_callbacks.hpp"
+#include "cutlass/epilogue/collective/xe_epilogue.hpp"
+#include "cutlass/epilogue/fusion/xe_callbacks.hpp"
 
 
 namespace cutlass::epilogue::collective {
