+++ include/cutlass/gemm/collective/collective_builder.hpp
@@ -40,6 +40,6 @@
 #include "cutlass/gemm/collective/builders/sm90_gmma_builder.inl"
 
 #if defined(SYCL_INTEL_TARGET)
-#include "cutlass/gemm/collective/builders/pvc_mma_builder.inl"
+#include "cutlass/gemm/collective/builders/xe_mma_builder.inl"
 #endif
 /////////////////////////////////////////////////////////////////////////////////////////////////
