+++ include/cutlass/gemm/kernel/gemm_universal.hpp
@@ -64,6 +64,6 @@ struct IsCutlass3ArrayKernel<ProblemShape, cute::void_t<typename ProblemShape::U
 #include "cutlass/gemm/kernel/sm90_gemm_array_tma_warpspecialized_cooperative.hpp"
 
 #if defined(SYCL_INTEL_TARGET)
-#include "cutlass/gemm/kernel/intel_pvc_gemm.hpp"
+#include "cutlass/gemm/kernel/xe_gemm.hpp"
 #endif
 ////////////////////////////////////////////////////////////////////////////////
